[native] Rename MarkdownLinkContext to MarkdownContext
Summary:
This is Step 1 of ENG-2072
(https://linear.app/comm/issue/ENG-2072/maintain-the-spoiler-state-upon-opening-textmessagetooltipmodal).
In this diff, we rename `MarkdownLinkContext` to just `MarkdownContext` in
order to (in a later step of this task) make this Context usable for the
spoilers and so we only need one Provider. Depends on D5383

Test Plan:
The app runs the same as it did prior to these changes, and
tested to see that link markdown behaves the same as before.

Reviewers: atul, ginsu, abosh

Reviewed By: atul, ginsu

Subscribers: ashoat, tomek, abosh

Differential Revision: https://phab.comm.dev/D5446