[services] Tunnelbroker - Changes to build the Tunnelbroker using `cargo` in a Docker
Summary:
This diff introduces changes to the Tunnelbroker's `Dockerfile` to change the build system from `cmake` to use `cargo` instead.
Also, deprecated cmake-related commands were removed from the `Dockerfile`.

This diff will also fix the CI (Docker) workflow fails for Tunnelbroker which are related to the Tunnelbroker's transition to Rust.

Related Linear task: [[ https://linear.app/comm/issue/ENG-2063/make-changes-to-tunnelbroker-docker-build-process-to-change-build | ENG-2063 ]]

**CI Notice:**
The CI will fail on this diff on the `Tunnelbroker Unit Tests (Nix)` until the D5461 where all the changes to CI build commands are made.
This stack will be** landed all in one** to prevent CI from failing on diffs out of this stack.

Test Plan:
1. The Tunnelbroker Docker CI build should pass successfully.

2. To test it manually:
Go to the Tunnelbroker service directory `services/tunnelbroker` and run the Docker compose build command `docker compose build tunnelbroker-server`.
The expected result is that the Docker image with the Tunnelbroker service is successfully built.

Reviewers: jon, varun, marcin, tomek

Reviewed By: jon, tomek

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5458