[web] Call searchUsers in chat composer
Summary:
[ENG-1988](https://linear.app/comm/issue/ENG-1988)
Adds a call to `searchUsers` in chat composer so a new chat can be created with users that are not in userStore.
Results from `searchUsers` are merged with userStore, and are treated as users that the viewer doesn't have a relationship with.
To enable chat creation, `forceAddMembers` is now set to `true` in `threadCreationResponder`.

Test Plan:
- Test if users that are not in `userStore` display in the search list
- Try selecting/ unselecting them and leaving chat composer when they are selected
- Try creating a new chat with multiple users that are not in `userStore`

Reviewers: tomek, atul, abosh

Subscribers: