[native] Refactor Spoiler component to MarkdownSpoiler
Summary: For consistency with the existing MarkdownLink component in
native, refactoring Spoiler to MarkdownSpoiler and putting it in the same
directory `native/markdown/...` makes the most sense. This is Step 2 in
ENG-2072
(https://linear.app/comm/issue/ENG-2072/maintain-the-spoiler-state-upon-opening-textmessagetooltipmodal).

Test Plan: Spoiler text still works as expected, and they are revealed on
click.

Reviewers: atul!, ginsu

Subscribers: