[services] Tunnelbroker - Move ffi from `main.rs` to the `cxx_bridge.rs` file
Summary:
This diff moves the FFI functions from CXX from `main.rs` to the `cxx_bridge.rs` file.
The reason for moving the FFI away from the `main.rs` to its own scope is that the `ffi::*` functions should be called from the gRPC server handlers (for example in D5487).

Linear task: [[ https://linear.app/comm/issue/ENG-2094/adding-dependencies-and-grpc-tonic-core-in-tunnelbroker-rust | ENG-2094 ]]

Test Plan: The service was successfully built with the CI process or can be built manually using the `cargo build` command in `services/tunnelbroker` directory.

Reviewers: jon, varun, marcin, tomek

Reviewed By: jon, tomek

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5486