[native] Added highlight effect to arrow background
Summary:
Solving this: https://linear.app/comm/issue/ENG-2045/add-highlight-effect-to-arrow-background-when-finger-is-on-thread-row

Reorganised `SidebarListModal`, `ChatThreadListSidebar` and `SidebarItem`, so that
`SidebarItem` doesn't hold button anymore, but rather is wrapped inside it.
Changed arrows SVG, so instead of short and long ones we have arrows extending in both ways and
short arrows that don't extend downwards, so they can finish list of thread sidebars.

{F208809}

Test Plan:
Build app in iOS and Android simulator.
Performed extensive visual testing if component is rendered correctly.

Reviewers: tomek, kamil, atul, bartek, ginsu

Reviewed By: tomek, kamil, bartek, ginsu

Subscribers: ginsu, abosh, bartek, atul, kamil, ashoat, tomek

Differential Revision: https://phab.comm.dev/D5491