Rename deviceIDTypes to deviceTypes, add comments
Summary: As suggested in comments on D4754, deviceIDTypes has been renamed to deviceTypes. The comments are needed because we have a logical dependency between these functions and variables that are not enforced by the code.

Test Plan: See with Redux DevTools that device ID is still present in Redux store on web

Reviewers: tomek, michal, kamil, bartek, jon

Reviewed By: tomek

Subscribers: ashoat, atul, abosh

Differential Revision: https://phab.comm.dev/D5301