[native] Introduce `useStaffCanSee()`
Summary:
Context: https://linear.app/comm/issue/ENG-2134/have-a-flag-that-indicates-whether-the-build-is-a-staff-release

Utility that tells us whether we're running a debug build (`__DEV__`) and/or are "staff" (`isStaff(state.currentUserInfo.id)`) and/or are on a staff release (`isStaffRelease`).

In subsequent diff will use this function where `isStaff || __DEV__` checks were used in the past.

Test Plan:
Logged values and they were as expected:

{F209770}

Reviewers: tomek, marcin, varun, ashoat

Reviewed By: ashoat

Subscribers: ashoat, abosh

Differential Revision: https://phab.comm.dev/D5498