[web] Change alert type in UserListItem
Summary:
Raised in D5518, it doesn't make sense for only one of `alertText` and `alertTitle` to be defined. This diff changes the type to make sure they are always defined together and updates the usages.

Depends on D5463

Test Plan: Check for flow errors. Check if thread composer works correctly on web and native

Reviewers: tomek, atul, abosh

Subscribers: ashoat