[services] Tunnelbroker - Removing of 'sessionID' from 'MessageToTunnelbroker'
Summary:
This diff removes the `sessionID` from the messages incoming to the Tunnelbroker. The purpose of this field was client authentication by providing the `sessionID` from the client. This method is not efficient because we will repeat the `sessionID` on every message instead of providing it once when the stream is opening.

Instead of repeating the `sessionID` on every message, we should use a gRPC canonical way for the authentication which is about using the [[ https://grpc.io/docs/what-is-grpc/core-concepts/#metadata | context metadata ]]. The full context is available in the Linear task.

Linear task: [[ https://linear.app/comm/issue/ENG-1359/remove-duplication-of-sessionid-in-every-tunnelbroker-message-and-use | ENG-1359 ]]

Test Plan: Services were successfully built, protobuf file was successfully compiled.

Reviewers: jon, tomek, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5525