[services] Tunnelbroker - Removing of emptiness check for notification token
Summary:
In this diff, we are removing the database item internal check for the device's notification token field.
We are fixing the bug where the database error occurs in the case when the device token in the database became empty and the server will return a database error when retrieving the token in the next client connection.

The field can be empty and the database check shouldn't fail on this. For more context please refer to the D5536.

Linear task: [[ https://linear.app/comm/issue/ENG-2060/implement-tunnelbroker-grpc-messagesstream-api-handler-in-rust | ENG-2060 ]]

Test Plan: This is a legacy C++ codebase change, the expected successful result is the successful building of the app and passing the CI checks.

Reviewers: jon, tomek

Reviewed By: jon, tomek

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5535