[web] Add spoiler rule to simpleMarkdownRules
Summary:
We add the rule to `web/markdown/rules.react.js` in order to parse content enclosed within the `|| ... ||` characters to be spoilers. This is the web implementation of the spoiler feature.

Linear: https://linear.app/comm/issue/ENG-2190/add-spoiler-rule-to-simplemarkdownrules-in-webmarkdownrulesreactjs

Test Plan:
Looked at the output in the inspect element tab on a specific message that has spoiler content. For now, until the next diff, the spoiler content is enclosed within `span` tags, signifying that the spoiler parse rule is working.

{F226270}

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: ashoat, tomek, abosh

Differential Revision: https://phab.comm.dev/D5546