[services][identity] introduce RPC to return the public key associated with a given user ID and device ID
Summary: keyservers need to be able to look up the public key associated with a given user and device ID in order to verify signed messages from users. this diff introduces a new RPC for this lookup.

Test Plan: successfully built the Identity service and native rust library, which consume this .proto file

Reviewers: ashoat, atul, tomek, jon

Reviewed By: ashoat, tomek

Subscribers: abosh

Differential Revision: https://phab.comm.dev/D5548