[web] Create a MarkdownSpoiler component for the web implementation to render when parsed
Summary:
We create a custom `MarkdownSpoiler` component to render spoiler content in-line, but to add additional functionality within the component in order to not dilute `rules.react.js`.

Linear:
https://linear.app/comm/issue/ENG-2194/create-a-markdownspoiler-component-for-the-web-implementation-to

Depends on D5546

Test Plan: The expected behavior is the same as the previous diff - the spoiler text still renders in a span.

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: ashoat, tomek, abosh

Differential Revision: https://phab.comm.dev/D5553