[services][identity] implement get_user_public_key service method
Summary:
this service method returns the public key for a given user/device combo if it's present. if it's not present, we return a not found error. if we encounter a different error, we handle it appropriately with the `handle_db_error` helper function.

Depends on D5569

Test Plan: ran the Identity service locally, and used Bloom to send RPCs to the service. tried with an invalid user ID, valid user ID and invalid device ID, and valid user and device ID, and confirmed that the expected response was received each time. Also tried running the service with invalid AWS credentials and saw the correct DB error in the logs.

Reviewers: tomek, bartek, jon

Reviewed By: tomek

Subscribers: ashoat, atul, abosh

Differential Revision: https://phab.comm.dev/D5570