[native] Deprecate react-native-dark-mode
Summary:
I'm working on [ENG-614](https://linear.app/comm/issue/ENG-614/update-to-react-native-070) to update React Native. The latest version of `react-native-dark-mode` had issues building on Android, and support for this functionality was introduced into the React Native core a while back anyways.

This diff deprecates `react-native-dark-mode` in favor of using the built-in `Appearance` module in React Native.

Test Plan: I tested this diff on the latest master and deployed it to my iOS device. I then set the iOS device to defer to the OS for dark mode, and swapped the OS settings and confirmed the app also swapped its setting

Reviewers: atul, tomek, bartek

Reviewed By: atul

Subscribers: abosh

Differential Revision: https://phab.comm.dev/D5605