[web] Extract thread-related selectors out of web/selectors/nav-selectors
Summary:
There is currently a "require" chain between `keyserver` and `web/markdown`, where it makes it impossible to use css files within the directory since `keyserver` is unable to handle it. Context:
https://linear.app/comm/issue/ENG-2206/extract-useonclickpendingsidebar-from-webselectorsnav-selectors#comment-31e14467. This means that in order to get around this, we need to break this "require" chain to
enable the use of `css` files in `MarkdownSpoiler`. This diff handles that - we separate the thread-like selectors out of `web/selectors/nav-selectors` into a new file, `web/selectors/thread-selectors`, and
fix the import statements around the codebase. This allows this chain to be broken and, after testing, `MarkdownSpoiler` importing a css file throws no error.

Depends on D5553

Test Plan:
Restart both `keyserver` and `web` with `yarn dev`, and there are no errors in the logs. Also manually checked each of the imports to make sure that they were all correct and up-to-date. Lastly, I
ran `flow` in the console within `web` to ensure that there were no flow-related issues that came out of this change.

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: ashoat, tomek, abosh

Differential Revision: https://phab.comm.dev/D5593