[web] Reveal the spoiler content on click within a message
Summary:
We add an `onClick` function to allow the spoiler content to be revealed on click. This `onClick` function then removes the `spoiler` class and animates the "reveal" effect to display the text content. The spoiler tags will not be re-rendered until the messages are unmounted.

https://linear.app/comm/issue/ENG-2192/reveal-the-spoiler-content-on-click-within-a-message

Depends on D5599

Test Plan:
Spoiler tags are revealed as expected, with a timed animation to fade the background and they get re-rendered after leaving and re-opening the chat.

{F231875}

Reviewers: atul, ginsu, ashoat

Reviewed By: atul, ginsu, ashoat

Subscribers: ashoat, tomek, abosh

Differential Revision: https://phab.comm.dev/D5601