[native] Avoid hardcoding path to rust_native_library lib.rs.h file
Summary: In D5480 we hardcoded a path to this that ended up changing in a later version of React Native. This diff updates the code so the path is passed directly from `corrosion-cxx.cmake` to where it's consumed.

Test Plan: I tested an Android debug build on both current `master` and my React Native upgrade branch and it worked

Reviewers: varun, jon, max

Reviewed By: varun

Subscribers: tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5604