[keyserver] Stop returning set of `deviceToken`s in `rescindPushNotifs(...)`
Summary:
We were previously using the set of `deviceToken`s returned by `rescindPushNotifs(...)` in `rescindAndUpdateBadgeCounts` to skip `updateBadgeCount(...)` for devices that had notifications rescinded. See D5606 for context.

Basically, we don't need the set of `deviceToken`s in `rescindAndUpdateBadgeCounts` and it isn't used anywhere else so it can be removed.

---

Depends on D5606

Test Plan:
Looked at other usages of `rescindPushNotifs(...)` and the return value was never used:
{F232108}

{F232109}

{F232110}

{F232111}

Reviewers: tomek, marcin, varun, ginsu, rohan, ashoat

Reviewed By: ashoat

Subscribers: ashoat, abosh

Differential Revision: https://phab.comm.dev/D5607