[services] Get rid of `return std::move(...)` in services codebase
Summary:
Context: https://linear.app/comm/issue/ENG-619/get-rid-of-stdmove-in-the-services-codebase

https://developers.redhat.com/blog/2019/04/12/understanding-when-not-to-stdmove-in-c
https://phabricator.ashoat.com/D2912?id=8614#inline-15820
https://phabricator.ashoat.com/D2876?id=8631#inline-15961

NOTE: We don't want to remove `std::move` altogether, just when we're returning so we don't interfere with return value optimization. We're still going to need to `std::move` `unique_ptrs` into lambdas and whatnot.

Test Plan: CI and close reading and previous discussion

Reviewers: tomek, marcin, kamil, inka, varun, max, jon, O2 Blocking Reviewers, ashoat

Reviewed By: max, O2 Blocking Reviewers, ashoat

Subscribers: ashoat, abosh

Differential Revision: https://phab.comm.dev/D5608