[native] Fix appearance of Inbox header in light mode
Summary:
Context: https://linear.app/comm/issue/ENG-1395/inbox-header-remains-dark-in-light-mode
https://linear.app/comm/issue/ENG-2052/re-introduce-light-mode-in-native

The `backgroundColor` of the Inbox header was previously hardcoded to a dark value that looked fine in dark mode but looked broken in light mode.

This diff sets the `backgroundColor` to `tabBarBackground` which will adapt based on the theme.

Test Plan:
Before:

{F232290}

After:

{F232291}

Things looks as expected.

Reviewers: tomek, marcin, ginsu, rohan, abosh, varun, O2 Blocking Reviewers, ashoat

Reviewed By: ginsu, rohan, O2 Blocking Reviewers, ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5610