[native] Fix appearance of `TabNavigator` in light mode
Summary:
Context: https://linear.app/comm/issue/ENG-2052/re-introduce-light-mode-in-native

The `backgroundColor` of `TabNavigator` was previously hardcoded to a dark value that looked fine in dark mode but looked broken in light mode.

This diff sets the `backgroundColor` to `tabBarBackground` which will adapt based on the theme.

Test Plan:
Before:

{F232316}

After:

{F232317}

Things look as expected.

Reviewers: tomek, marcin, ginsu, rohan, abosh, varun, O2 Blocking Reviewers, ashoat

Reviewed By: ginsu, O2 Blocking Reviewers, ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D5611