Display an alert when sensitive data are purged
Summary: This differential alerts: developer, staff members and users on a staff release when sensitive data (currently database and data in secure store) are removed due to change in credentials of currently logged-in user.

Test Plan: Build the app, log out, create new account then delete it. in both cases the alert should be displayed.

Reviewers: atul, ashoat

Reviewed By: atul

Subscribers: tomek, abosh

Differential Revision: https://phab.comm.dev/D5626