[native] fixed tooltip triangle color
Summary:
Found a small ui bug as I was working with the tooltips where the triangles had a different color than `styles.items` and was dark in "light mode".

Also as a side note, I think the color for the tooltip is a little too dark and it's hard to see the triangles. I didn't change it because the focus of this diff is to unify the color of the triangles, but just something we should think about especially now that we have some designers on the team

---

Linear Task: [[ https://linear.app/comm/issue/ENG-2052/re-introduce-light-mode-in-native | ENG-2052 ]]

Test Plan:
Please see the screenshots to see the changes I made

Before:
Dark Mode:
{F241099}

Light Mode:
{F241100}

After:
Dark Mode:
{F241102}

Light Mode:
{F241104}

Reviewers: atul, rohan

Reviewed By: atul

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5628