[services] Tunnelbroker - Expose config to the Rust
Summary:
In this diff, we are exposing a function to get the parameter from the config file from the C++ codebase to the Rust codebase as a `getConfigParameter` function.
This function will be used in the Rust codebase diff D5641 and forward.

Linear task: [[ https://linear.app/comm/issue/ENG-1893/rust-part-code-small-refactors | ENG-1893 ]]

Test Plan:
1. Service is successfully built.
2. The diff D5641 where the function is used the service is successfully built.

Reviewers: jon, varun

Reviewed By: jon

Subscribers: ashoat, tomek, atul, abosh

Differential Revision: https://phab.comm.dev/D5639